adventure.vueに統合
